Fix 6 Codex review findings on the message library foundation
- Migration: selection_score was NUMERIC but the ORM declares Float
  (schema_audit.py expects double precision); deploy's post-migration
  audit would have failed. Changed to DOUBLE PRECISION. Verified with
  scripts/schema_audit.py against a fresh database: passes clean.
- created_by had no length bound, so an oversized value crashed
  flush() instead of returning the documented fail-closed result.
- set_key is a permanent identifier (same convention as Identity
  Registry's registry_key) but carries an unconditional UNIQUE
  constraint; recreating a deleted set_key crashed flush() instead of
  being rejected cleanly.
- A malformed placeholder (e.g. {{first-name}}, hyphen instead of
  underscore) didn't match the strict token pattern, so it silently
  passed creation validation and would have rendered as literal,
  broken template text.
- record_variant_usage() didn't verify the variant actually belongs
  to the given message_set_id, so a mismatched pair silently
  corrupted the per-set analytics trail.
- delete_message_set() only tombstoned the parent; child variants
  stayed enabled and independently fetchable/renderable, so deleting
  a set didn't actually retire its content. Now cascades the
  tombstone to every active child variant.

5 new regression tests added (24 total in this file, all passing).
Re-verified full suite (821 passed, 76.42% coverage), DB leakage
(clean), P0 outbound-authority matrix (71 passed, unchanged), and the
full 001-034 migration chain against a fresh database.

+++ bot_core/migrations/034_outbound_message_library.sql
@@ -74,7 +74,7 @@ CREATE TABLE IF NOT EXISTS outbound_variant_usage (
     message_set_id BIGINT NOT NULL REFERENCES outbound_message_sets(id) ON DELETE CASCADE,
     variant_id BIGINT NOT NULL REFERENCES outbound_message_variants(id) ON DELETE CASCADE,
     outbound_queue_id BIGINT NULL REFERENCES outbound_queue(id) ON DELETE SET NULL,
-    selection_score NUMERIC NULL,
+    selection_score DOUBLE PRECISION NULL,
     selection_reason TEXT NULL,
     source_automation VARCHAR(120) NULL,
     send_result VARCHAR(20) NOT NULL DEFAULT 'pending',
